Moved output layout to new file, fixed constructor
It was not checking the null pointer, oops!
