+++ src/lib.rs
@@ -54,7 +54,8 @@ pub use self::types::area::*;
 pub use self::types::cursor::*;
 pub use self::types::input_device::*;
 pub use self::types::keyboard::*;
-pub use self::types::output::*;
+pub use self::types::output::output::*;
+pub use self::types::output::output_layout::*;
 pub use self::types::pointer::*;
 pub use self::types::seat::*;
 pub use self::types::surface::*;
